@@ -2,7 +2,7 @@
 
     <PropertyGroup>
         <!-- Multi-target for compatibility -->
-        <TargetFrameworks>net8.0;net472</TargetFrameworks>
+        <TargetFrameworks>net10.0;net9.0;net8.0;net472</TargetFrameworks>
         <GenerateDocumentationFile>false</GenerateDocumentationFile>
         <Description>
             MSBuild SDK for DotNetDocs projects (.docsproj) - provides a clean way to include documentation in .NET
@@ -35,7 +35,8 @@
         <PackageReference Include="Microsoft.Build.Framework" Version="17.*" PrivateAssets="all" />
         <PackageReference Include="Microsoft.Build.Utilities.Core" Version="17.*" PrivateAssets="all" ExcludeAssets="runtime" />
         <PackageReference Include="System.Collections.Immutable" Version="9.*" PrivateAssets="all" />
-        
+        <PackageReference Include="Microsoft.CodeAnalysis.CSharp.Workspaces" Version="4.*" />
+
         <!-- Include Microsoft.Build.NoTargets for bundling its SDK files -->
         <PackageReference Include="Microsoft.Build.NoTargets"
                           Version="3.7.0"
@@ -44,6 +45,14 @@
     </ItemGroup>
 
     <!-- Framework-specific dependencies -->
+    <ItemGroup Condition="'$(TargetFramework)' == 'net10.0'">
+        <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="10.*-*" />
+    </ItemGroup>
+
+    <ItemGroup Condition="'$(TargetFramework)' == 'net9.0'">
+        <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="9.*" />
+    </ItemGroup>
+
     <ItemGroup Condition="'$(TargetFramework)' == 'net8.0'">
         <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="9.*" />
     </ItemGroup>
